feat: support SVG format
