+++ src/addon/__init__.py
@@ -36,7 +36,6 @@ dll: ctypes.CDLL
 
 classes: list[type] = [
     # menus
-    menus.VIEW3D_MT_Space_Import_SVG,
     menus.VIEW3D_MT_Space_Import_USD,
     menus.VIEW3D_MT_Space_Import_USDA,
     menus.VIEW3D_MT_Space_Import_USDC,
