+++ src/addon/formats/__init__.py
@@ -15,6 +15,7 @@ from . import obj_legacy
 from . import ply
 from . import stl
 from . import stl_legacy
+from . import svg
 from . import vrm
 
 CLASSES: list[type] = []
@@ -32,4 +33,5 @@ CLASSES.extend(glb.OPERATORS)
 CLASSES.extend(obj.OPERATORS)
 CLASSES.extend(ply.OPERATORS)
 CLASSES.extend(stl.OPERATORS)
+CLASSES.extend(svg.OPERATORS)
 CLASSES.extend(vrm.OPERATORS)
